fix: update about page content to reduce link complexity
