--- app/site-config/about/about__page.ts
@@ -1,30 +1,0 @@
-import type { ContentBlock } from "../types";
-
-export type AboutPageBody = {
-  body: ContentBlock[];
-};
-
-export const ABOUT_PAGE_BODY: AboutPageBody = {
-  body: [
-    {
-      type: "text",
-      heading: "Advancing Science for Disaster Resilience",
-      headingLevel: "h2",
-      paragraphs: [
-        "The NASA Disasters Program puts Earth science to work for those who make critical decisions before, during, and after disasters. We translate NASA’s unmatched view of Earth from space into actionable insights, helping emergency managers, government agencies, and industry partners prepare for high-impact hazards, respond effectively when disasters strike, and recover more fully in their aftermath. From hurricanes and volcanoes to floods and earthquakes, we use NASA’s data, tools, and expertise to build resilience in communities across the U.S. and around the world.",
-      ],
-    },
-    {
-      type: "link",
-      label: "Contact Our Team",
-      href: "mailto:hq-em-disasters@mail.nasa.gov",
-    },
-    {
-      type: "link",
-      label: "Sign up for the NASA Disasters Community Newsletter",
-      href: "https://lp.constantcontactpages.com/sl/ICIOyJI",
-      target: "_blank",
-      rel: "noopener noreferrer",
-    },
-  ],
-};
